AI-Interview/requirements.txt
google-generativeai
PyPDF2
python-dotenv
speechrecognition
gTTS
pygame